KC_UNDS → NAV_UND for switching to NAV from SYM
When I was trying to quickly move a certain amount of lines up or down
in vim, I was noticing that the nav layer wasn't actually getting
activated so I often ended up with unintentional `5u` or `3e`.

The solution would be to turn the underscore key into a layer tap but
since KC_UNDS isn't a basic keycode, a trick had to be pulled.